Refactor templates: Rule and Templatable and products
Move make_items_product_specific method from Rule into common module
as static function along with GLOBAL_REFERENCES (constants). This
would allow us to reuse code beyond Rule component. Now Templatable
can normalize/productify template variables on its own.

Clean up duplicate definition of add_sub_element function.
