+++ ssg/entities/common.py
@@ -15,7 +15,46 @@ from ..constants import (
     XCCDF_REFINABLE_PROPERTIES,
     XCCDF12_NS,
     OSCAP_VALUE,
+    GLOBAL_REFERENCES
 )
+
+
+def make_items_product_specific(items_dict, product_suffix, allow_overwrites=False):
+    new_items = dict()
+    for full_label, value in items_dict.items():
+        if "@" not in full_label and full_label not in new_items:
+            new_items[full_label] = value
+            continue
+
+        label = full_label.split("@")[0]
+
+        # This test should occur before matching product_suffix with the product qualifier
+        # present in the reference, so it catches problems even for products that are not
+        # being built at the moment
+        if label in GLOBAL_REFERENCES:
+            msg = (
+                "You cannot use product-qualified for the '{item_u}' reference. "
+                "Please remove the product-qualifier and merge values with the "
+                "existing reference if there is any. Original line: {item_q}: {value_q}"
+                .format(item_u=label, item_q=full_label, value_q=value)
+            )
+            raise ValueError(msg)
+
+        if not full_label.endswith(product_suffix):
+            continue
+
+        if label in items_dict and not allow_overwrites and value != items_dict[label]:
+            msg = (
+                "There is a product-qualified '{item_q}' item, "
+                "but also an unqualified '{item_u}' item "
+                "and those two differ in value - "
+                "'{value_q}' vs '{value_u}' respectively."
+                .format(item_q=full_label, item_u=label,
+                        value_q=value, value_u=items_dict[label])
+            )
+            raise ValueError(msg)
+        new_items[label] = value
+    return new_items
 
 
 def add_sub_element(parent, tag, ns, data):
@@ -323,3 +362,19 @@ class Templatable(object):
 
     def is_templated(self):
         return isinstance(self.template, dict)
+
+    def make_template_product_specific(self, product):
+        if not self.is_templated():
+            return
+
+        product_suffix = "@{0}".format(product)
+
+        not_specific_vars = self.template.get("vars", dict())
+        specific_vars = make_items_product_specific(
+            not_specific_vars, product_suffix, True)
+        self.template["vars"] = specific_vars
+
+        not_specific_backends = self.template.get("backends", dict())
+        specific_backends = make_items_product_specific(
+            not_specific_backends, product_suffix, True)
+        self.template["backends"] = specific_backends
